+++ app/controllers/carto/api/user_presenter.rb
@@ -123,6 +123,7 @@ module Carto
           account_type_display_name: plan_name(@user.account_type),
           table_quota: @user.table_quota,
           public_map_quota: @user.public_map_quota,
+          public_dataset_quota: @user.public_dataset_quota,
           private_map_quota: @user.private_map_quota,
           regular_api_key_quota: @user.regular_api_key_quota,
           table_count: @user.table_count,