+++ spec/requests/carto/api/visualizations_controller_spec_part_2.rb
@@ -1354,6 +1354,7 @@ describe Carto::Api::VisualizationsController do
       after(:each) do
         @carto_user1.private_maps_enabled = false
         @carto_user1.public_map_quota = nil
+        @carto_user1.public_dataset_quota = nil
         @carto_user1.private_map_quota = nil
         @carto_user1.save
       end
@@ -1388,6 +1389,36 @@ describe Carto::Api::VisualizationsController do
 
         last_response.status.should == 403
         last_response.body.should =~ /public map quota/
+      end
+
+      it 'returns a 200 response when making a dataset public with enough quota' do
+        @carto_user1.public_dataset_quota = nil
+        @carto_user1.save
+        visualization = FactoryGirl.create(:carto_table_visualization, user: @carto_user1,
+                                           privacy: Carto::Visualization::PRIVACY_PRIVATE)
+
+        request_params = { user_domain: @carto_user1.username, api_key: @carto_user1.api_key, id: visualization.id }
+        put api_v1_visualizations_update_url(request_params),
+            { id: visualization.id, privacy: CartoDB::Visualization::Member::PRIVACY_PUBLIC }.to_json,
+            @headers
+
+        last_response.status.should == 200
+        visualization.destroy!
+      end
+
+      it 'returns a 403 response when making a dataset public without enough quota' do
+        @carto_user1.public_dataset_quota = 0
+        @carto_user1.save
+        visualization = FactoryGirl.create(:carto_table_visualization, user: @carto_user1,
+                                           privacy: Carto::Visualization::PRIVACY_PRIVATE)
+
+        request_params = { user_domain: @carto_user1.username, api_key: @carto_user1.api_key, id: visualization.id }
+        put api_v1_visualizations_update_url(request_params),
+            { id: visualization.id, privacy: CartoDB::Visualization::Member::PRIVACY_PUBLIC }.to_json,
+            @headers
+
+        last_response.status.should == 403
+        last_response.body.should =~ /public dataset quota/
       end
 
       it 'returns a 200 response when making a map private with enough quota' do