+++ spec/services/carto/visualizations_export_service_2_spec.rb
@@ -1616,6 +1616,21 @@ describe Carto::VisualizationsExportService2 do
         @user2.unstub(:public_map_quota)
       end
 
+      it 'raises an unauthorized error when the import is over public dataset quota' do
+        @visualization.privacy = 'password'
+        @visualization.user_id = @user2.id
+        @visualization.password = 'super_secure_secret'
+        @visualization.save!
+
+        @user2.stubs(:public_dataset_quota).returns(1)
+        exported_string = export_service.export_visualization_json_string(@visualization.id, @user, with_password: true)
+        built_viz = export_service.build_visualization_from_json_export(exported_string)
+        expect {
+          Carto::VisualizationsExportPersistenceService.new.save_import(@user2, built_viz)
+        }.to raise_error(Carto::UnauthorizedError)
+        @user2.unstub(:public_dataset_quota)
+      end
+
       it 'raises an unauthorized error when the import is over private map quota' do
         @visualization.privacy = 'private'
         @visualization.user_id = @user2.id